requirements.txt
numpy~=1.25.0
tensorflow~=2.17.0
PyYAML~=6.0.0

future~=0.18.3
Pillow~=9.4.0
setuptools~=68.2.0
matplotlib~=3.7.0
seaborn~=0.12.0